Avoid duplicate MODULEPATH for spider in CUDA-on-top-of-MPI modules
CUDA-on-MPI adds both the MPI and Compiler CUDA MODULEPATHs, and
CUDA-on-compiler only the Compiler one. This confuses `module spider`.
So we need to guard it via a hook, and the MNS should only set the
MPI MODULEPATH (the MNS isn't smart enough for this syntax, the
hook needs to add it).

```
if (mode() ~= "spider") then
    prepend_path("MODULEPATH", pathJoin("/cvmfs/soft.computecanada.ca/easybuild/modules/2023/x86-64-v3/CUDA/gcc12/cuda12.2")
    if isDir(pathJoin(os.getenv("HOME"), ".local/easybuild/modules/2023/x86-64-v3/CUDA/gcc12/cuda12.2")) then
        prepend_path("MODULEPATH", pathJoin(os.getenv("HOME"), ".local/easybuild/modules/2023/x86-64-v3/CUDA/gcc12/cuda12.2"))
    end
end
```